refort: axisPointer tooltip.
